Reordered code a bit too avoid copy/paste code.
